Move realpath modelling to header file (no code changes)
